增加了消费趋势页面

+++ app/src/main/java/com/duoji/app/ui/home/HomeViewModel.kt
@@ -159,7 +159,7 @@ class HomeViewModel : ViewModel() {
 
         val daysBetween = ChronoUnit.DAYS.between(startDate, endDate).toInt()
         val points = (0..daysBetween).map { days ->
-            val date = startDate.plusDays(days)
+            val date = startDate.plusDays(days.toLong())
             DailyExpensePoint(
                 date = date,
                 amount = expensesByDate[date] ?: 0.0
